+++ main.py
@@ -1,5 +1,5 @@
 from system import account
 
-print("RUNE OS ALPHA 3.5.1")
+print("RUNE OS ALPHA 3.5.3")
 print("Booting up...")
 account.login()